# [Kotlin/Native] Reduce the overhead of EnterFrame/LeaveFrame by a factor of 4x on Windows