This reduces the overhead of EnterFrame/LeaveFrame by a factor of 4x

* https://kotlinlang.slack.com/archives/C3SGXARS6/p1619349974244300
* https://github.com/korlibs/kotlin-native-performance-experiment

To try this I have backported this change to Kotlin 1.4.32 that doesn't remove EnterFrame/LeaveFrame in most cases and run the sample from here: https://github.com/korlibs/kotlin-native-performance-experiment.

Without the patch:

```kotlin
Executing INLINED...
Executed INLINED sixty frames in 865ms
Executing NON-INLINED-BUT-CONST...
Executed NON-INLINED-BUT-CONST sixty frames in 38.7s
Executing NON-INLINED...
Executed NON-INLINED sixty frames in 63.3s
```

Patched:

```kotlin
Executing INLINED...
Executed INLINED sixty frames in 864ms
Executing NON-INLINED-BUT-CONST...
Executed NON-INLINED-BUT-CONST sixty frames in 10.3s
Executing NON-INLINED...
Executed NON-INLINED sixty frames in 14.4s
```

LLVM thread locals on windows uses this implementation <https://github.com/llvm-mirror/compiler-rt/blob/69445f095c22aac2388f939bedebf224a6efcdaf/lib/builtins/emutls.c#L378> which much slower than using the WINDOWS APIs: <https://docs.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-tlsalloc>