@SvyatoslavScherbina 2021-06-04
Thank you for the contribution.

Sorry for the delay. We've needed some time to make the decision, including whether we should remove `-femulated-tls` from `konan.properties` (which would make this PR unnecessary). Apparently we shouldn't, because otherwise https://youtrack.jetbrains.com/issue/KT-27654 would be triggerred again.

Could you please also add a note here: https://github.com/JetBrains/kotlin/blob/f5d656ec4fe80ac2a867673a2b982f46b3c6dbbf/kotlin-native/konan/konan.properties#L872
about using `TlsAlloc` to mitigate the performance impact of `-femulated-tls`. This would help us to remember reverting it back to `THREAD_LOCAL_VARIABLE` if we decide to remove `-femulated-tls` later.