@soywiz 2021-06-11
Okay. A small update. I have performed this spike:

```
// g++ -shared -o dll.dll dll.cpp && g++ -o exe.exe exe.cpp && exe

// dll.cpp
#include <stdio.h>
#include <windows.h>

static void dllIndexUnalloc();
static DWORD dllIndexAlloc();

static DWORD globalTlsIndex = dllIndexAlloc();

static void dllIndexUnalloc() {
    printf("dllIndexUnalloc: %d\n", globalTlsIndex);
    TlsFree(globalTlsIndex);
}

static DWORD dllIndexAlloc() {
    DWORD globalTlsIndex = TlsAlloc();
    printf("dllIndexAlloc: %d\n", globalTlsIndex);
    if (globalTlsIndex == TLS_OUT_OF_INDEXES) abort();
    atexit(dllIndexUnalloc);
    return globalTlsIndex;
}

extern "C" {
    __declspec(dllexport) void print_dll_index() {
        printf("print_dll_index: %d\n", globalTlsIndex);
    }
}

// exe.cpp
#include <stdio.h>
#include <windows.h>

typedef void (*print_dll_index_callback)();

int main() {
    for (int n = 0; n < 10; n++) {
        HMODULE dll = LoadLibrary("dll.dll");
        print_dll_index_callback print_dll_index = (print_dll_index_callback)GetProcAddress(dll, "print_dll_index");
        printf("%p\n", dll);
        print_dll_index();
        FreeLibrary(dll);
    }
}
```

It seems that `atexit` works with `DllMain` internally and detects dll unloading. So it seems to work properly.

`TlsAlloc` seems to have 64 slots at least and 1088 in modern Windows versions.