@SvyatoslavScherbina 2021-06-15
Please don't forget to [add a comment](https://github.com/JetBrains/kotlin/pull/4339#issuecomment-854452176) to `konan.properties`.