@soywiz 2021-07-31
Sorry for the huge delay, I have been extra busy and totally forgot about this, sorry.

A few things I remember about this:
* Right now `emutls` leaks handlers too when you load and unload DLLs (I tried that), though that's pretty uncommon thing to do. So the `TlsFree` call is an improvement over the current implementation, but shouldn't be necessary.
* Do this properly would require to split the runtime in two, since the `#define` happens compiling the runtime, and not on the final program. Doing otherwise would improve the plain performance on Windows, but would also reduce the performance on our game engine, since we are disabling `emutls` in our case, and that could probably affect other people doing the same trick you suggested on slack.

I can try to implement the runtime split if you can give me some indications on how to do so, but I'm not sure that is worth.
As an alternative, in the case you think this idea is worth and not going to be deprecated, the code is here, feel free to use it 👍But please, keep supporting disabling `emutls` so we can continue benefiting from a performance comparable to Linux and Mac on Windows.

Thanks!