@SvyatoslavScherbina 2021-08-04
> Right now `emutls` leaks handlers too when you load and unload DLLs (I tried that), though that's pretty uncommon thing to do. So the `TlsFree` call is an improvement over the current implementation, but shouldn't be necessary.

Fair enough.

> Do this properly would require to split the runtime in two, since the `#define` happens compiling the runtime, and not on the final program. Doing otherwise would improve the plain performance on Windows, but would also reduce the performance on our game engine, since we are disabling `emutls` in our case, and that could probably affect other people doing the same trick you suggested on slack.

Makes sense.

> I can try to implement the runtime split if you can give me some indications on how to do so, but I'm not sure that is worth.

Sounds complicated, TBH. In particular, this would interfere with compiler caches.

Also, we are considering removing `-femulated-tls` in 1.6. See https://youtrack.jetbrains.com/issue/KT-47605
So let's keep things as is.

Anyway, thank you for the investigation and the PR!
This also motivated us to take a closer look at the problem and make relevant changes.